- Fixed Auditing not adding ID of users. - Added new Exception with corresponding ProblemDetail, to prohibid one user to modify another user data, and modified JsonContent service to check for current user. - Starting to remove custom auth in FilterChain and use instead method auth validation - Other minor changes

+++ src/main/java/com/duberlyguarnizo/dummyjson/jsoncontent/JsonContentService.java
@@ -1,12 +1,15 @@
 package com.duberlyguarnizo.dummyjson.jsoncontent;
 
+import com.duberlyguarnizo.dummyjson.Auditing.CustomAuditorAware;
 import com.duberlyguarnizo.dummyjson.exceptions.IdNotFoundException;
+import com.duberlyguarnizo.dummyjson.exceptions.NotOwnedObjectException;
 import com.duberlyguarnizo.dummyjson.exceptions.RepositoryException;
 import com.duberlyguarnizo.dummyjson.jsoncontent.dto.JsonContentBasicDto;
 import com.duberlyguarnizo.dummyjson.jsoncontent.dto.JsonContentCreationDto;
 import com.duberlyguarnizo.dummyjson.jsoncontent.dto.JsonContentDetailDto;
 import com.duberlyguarnizo.dummyjson.jsoncontent.dto.JsonContentMapper;
 import jakarta.validation.Valid;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,10 +18,12 @@ import java.util.List;
 public class JsonContentService {
     private final JsonContentRepository repository;
     private final JsonContentMapper mapper;
+    private final CustomAuditorAware auditorAware;
 
-    public JsonContentService(JsonContentRepository repository, JsonContentMapper mapper) {
+    public JsonContentService(JsonContentRepository repository, JsonContentMapper mapper, CustomAuditorAware auditorAware) {
         this.repository = repository;
         this.mapper = mapper;
+        this.auditorAware = auditorAware;
     }
 
     //CRUD
@@ -31,8 +36,17 @@ public class JsonContentService {
         }
     }
 
-    public List<JsonContentBasicDto> getAll() {
-        return repository.findAll().stream().map(mapper::toBasicDto).toList();
+    public List<JsonContentBasicDto> getAllByCurrentUser() {
+        var currentAuditor = auditorAware.getCurrentAuditor();
+        if (currentAuditor.isEmpty()) {
+            throw new AccessDeniedException("You do not have the permissions to list this resource.");
+        } else {
+            Long currentUserId = currentAuditor.get();
+            return repository.findAllByCreatedBy(currentUserId)
+                    .stream()
+                    .map(mapper::toBasicDto)
+                    .toList();
+        }
     }
 
     public Long create(@Valid JsonContentCreationDto jsonDto) {
@@ -46,22 +60,43 @@ public class JsonContentService {
     }
 
     public void update(@Valid JsonContentCreationDto jsonDto) {
-        var json = repository.findById(jsonDto.getId());
-        if (json.isEmpty()) {
-            throw new IdNotFoundException("No JSON content found with id: " + jsonDto.getId() + " in the database");
+        //TODO: refactor to use @POSTAUTHORIZE
+        var currentAuditor = auditorAware.getCurrentAuditor();
+        if (currentAuditor.isEmpty()) {
+            throw new AccessDeniedException("You do not have the permissions to update this resource.");
         } else {
-            JsonContent jsonObject = json.get();
-            var updatedJson = mapper.partialUpdate(jsonDto, jsonObject);
-            repository.save(updatedJson);
+            var json = repository.findById(jsonDto.getId());
+            if (json.isEmpty()) {
+                throw new IdNotFoundException("No JSON content found with id: " + jsonDto.getId() + " in the database");
+            } else {
+                Long currentUserId = currentAuditor.get();
+                JsonContent jsonObject = json.get();
+                if (jsonObject.getCreatedBy() != currentUserId) {
+                    throw new NotOwnedObjectException("You do not have the permissions to update this resource.");
+                }
+                var updatedJson = mapper.partialUpdate(jsonDto, jsonObject);
+                repository.save(updatedJson);
+            }
         }
     }
 
     public void delete(Long id) {
-        var json = repository.findById(id);
-        if (json.isEmpty()) {
-            throw new IdNotFoundException("No JSON content found with id: " + id + " in the database");
+        var currentAuditor = auditorAware.getCurrentAuditor();
+        if (currentAuditor.isEmpty()) {
+            throw new AccessDeniedException("You do not have the permissions to delete this resource.");
         } else {
-            repository.deleteById(id);
+            var json = repository.findById(id);
+            if (json.isEmpty()) {
+                throw new IdNotFoundException("No JSON content found with id: " + id + " in the database");
+            } else {
+                Long currentUserId = currentAuditor.get();
+                JsonContent jsonObject = json.get();
+                if (jsonObject.getCreatedBy() != currentUserId) {
+                    throw new NotOwnedObjectException("You do not have the permissions to delete this resource.");
+                }
+                repository.deleteById(id);
+            }
         }
     }
 }
+
